Make it possible to return a C++ `ShardedDeviceArray`.
This is a *breaking* change for the next Jax release with a new `jaxlib` (depending on what will be shipped in the next release or not).

- Add `make_sharded_device_array` to be used within JAX and for hackers that need to construct SDA objects.
- Make sure the C++ object is valid by
  (a) extending `DeviceArrayBase` (done in Python), as it brings a bunch of methods and enable `isinstance(x, DeviceArray)`
  (b) Adding the same methods as the Python SDA.

All tests in JAX pass when we set `_USE_EXPERIMENTAL_CPP_SDA=True`, but the switch will happen in its own CL.

Similarly to JAX.jit, for which we have a C++ `DeviceArray` and a Python `_DeviceArray`, we will introduce 2 objects for ShardedDeviceArray, with the Python object only for for JAX extensions not compatible with the C++ object (e.g. Cloud TPU).

PiperOrigin-RevId: 388190299